GDE-26: Merge the UIKit build into the theme json.
- Updates the theme package.json so that it includes the uikit modules.
- Updates the README.md file so that it includes some information on the uilkit
- Adds a uikit:refresh npm command that will help rebuild the uikit.
- Updates _config.scss to include all configuration variables used by uikit.
- Updates gulp tasks:
-- Adds a configuration for scss linting so we can exclude directories.
-- Updates css lint gulp task to use the new configuration value.
- Adds a local .gitignore to the STARTERKIT theme.

We're explictly excluding the uikit from the Cog linting standards as they are not compatible. We will rely on the testing that UI Kit does to ensure that it adheres to best practises dictated by that project but will leave Cog standards for ES Linting and CSS Linting for custom CSS+JS added by the theme developer.